agents/requirements.txt
langgraph>=0.2.0
anthropic>=0.34.0
openai>=1.40.0
python-dotenv>=1.0.0
asyncio
